@@ -9,8 +9,8 @@ plugins {
     id 'org.asciidoctor.convert' version '1.5.6'
 }
 
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
+sourceCompatibility = JavaVersion.VERSION_1_9
+targetCompatibility = JavaVersion.VERSION_1_9
 
 repositories {
     mavenCentral()
@@ -33,6 +33,11 @@ sourceSets {
 }
 
 dependencies {
+    implementation group: 'javax.xml.bind', name: 'jaxb-api', version: '2.2.8'
+    implementation group: 'com.sun.xml.bind', name: 'jaxb-impl', version: '2.3.0'
+    implementation group: 'com.sun.xml.bind', name: 'jaxb-core', version: '2.3.0'
+    implementation group: 'javax.activation', name: 'activation', version: '1.1.1'
+
     testImplementation group: 'junit', name: 'junit', version: '4.12'
 }
 
